Python_programs/requirements_chatbot/requirements.txt
python-dotenv==1.0.0
questionary==2.0.1
pyyaml==6.0.1
jinja2==3.1.2 